chore: merge후 수정

+++ phoncnic/src/main/java/ds/com/phoncnic/entity/Follow.java
@@ -22,7 +22,7 @@ public class Follow {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int fno;
+    private Long fno;
 
     @ManyToOne
     private Member follower;
